+++ server.js
@@ -9,6 +9,7 @@ import questionRoutes from './routes/question.js';
 import rankingRoutes from './routes/ranking.js';
 import userRoutes from './routes/user.js';
 import ratingRoutes from './routes/rating.js';
+import inviteRoutes from './routes/invite.js';
 import dotenv from 'dotenv';
 dotenv.config();
 import { v2 as cloudinary } from 'cloudinary';
@@ -40,6 +41,7 @@ app.use('/api/question', questionRoutes);
 app.use('/api/ranking', rankingRoutes);
 app.use('/api/user', userRoutes);
 app.use('/api/rating', ratingRoutes);
+app.use('/api/invites', inviteRoutes);
 // Basic route
 app.get('/', (req, res) => {
     res.json({